LayoutTests/fast/frames/seamless/seamless-min-max-expected.txt

PASS window.getComputedStyle(iframe1).width is "300px"
PASS window.getComputedStyle(iframe1).height is "300px"
FAIL window.getComputedStyle(iframe2).width should be 50px. Was 200px.
FAIL window.getComputedStyle(iframe2).height should be 50px. Was 0px.
